Refactored code to comply with SOLID
